refactor: serviceモジュールに端末に依存する機能を移動し、dataモジュールはデータの永続化に注力するように変更
