+++ core/service/src/main/AndroidManifest.xml
@@ -2,7 +2,7 @@
 
     <application>
         <service
-            android:name=".notification.NotificationListener"
+            android:name=".service.notification.NotificationListener"
             android:permission="android.permission.BIND_NOTIFICATION_LISTENER_SERVICE"
             android:exported="true">
 
@@ -12,7 +12,7 @@
         </service>
 
         <service
-            android:name=".wallpaper.UnityWallpaperService"
+            android:name=".service.wallpaper.UnityWallpaperService"
             android:permission="android.permission.BIND_WALLPAPER"
             android:exported="true">
