+++ core/service/src/main/kotlin/io/github/kei_1111/withmo/core/service/di/PermissionModule.kt
@@ -1,4 +1,4 @@
-package io.github.kei_1111.withmo.core.data.di
+package io.github.kei_1111.withmo.core.service.di
 
 import android.content.Context
 import dagger.Module
@@ -6,8 +6,8 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
-import io.github.kei_1111.withmo.core.data.permission.PermissionCheckerImpl
 import io.github.kei_1111.withmo.core.domain.permission.PermissionChecker
+import io.github.kei_1111.withmo.core.service.permission.PermissionCheckerImpl
 import javax.inject.Singleton
 
 @Module
